Refactor "today" determination
Instead of using `lazy_static` to declare the `TODAY` static ref, this
splits things up so that `main` defaults to the current date but the new
`print` function accepts a substitute "current date".

This allows the acceptance tests to be refactored to be much much closer
to reality (e.g. they actually go through the args parsing).